server/requirements.txt
backports.weakref==1.0rc1
BeautifulSoup==3.2.1
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.4
django-cors-headers==2.1.0
django-facebook==6.0.3
django-picklefield==1.0.0
djangorestframework==3.6.3
funcsigs==1.0.2
gunicorn==19.7.1
html5lib==0.9999999
idna==2.5
Markdown==2.6.8
mock==2.0.0
numpy==1.13.1
olefile==0.44
pbr==3.1.1
Pillow==4.2.1
protobuf==3.3.0
psycopg2==2.7.3
pytz==2017.2
requests==2.18.3
scikit-learn==0.19.0
scipy==0.19.1
six==1.10.0
sklearn==0.0
tensorflow==1.2.1
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12.2
whitenoise==3.3.0
